+++ pet-shop-dto/src/main/java/by/kachanov/shop/dto/Order.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonManagedReference;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
@@ -16,7 +17,7 @@ public class Order {
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "pk")
     @SequenceGenerator(name = "pk", sequenceName = "pk_seq", allocationSize = 1)
-    private BigDecimal id;
+    private BigInteger id;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
@@ -28,11 +29,11 @@ public class Order {
     @OneToMany(mappedBy = "order", cascade = CascadeType.ALL)
     private Set<OrderPosition> orderPositions = new HashSet<>();
 
-    public BigDecimal getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
