+++ pet-shop-repository/pet-shop-repository-api/src/main/java/by/kachanov/shop/repository/OrderRepository.java
@@ -3,14 +3,14 @@ package by.kachanov.shop.repository;
 import by.kachanov.shop.dto.Order;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface OrderRepository {
 
     void saveOrder(Order order);
 
-    Order getOrder(BigDecimal orderId);
+    Order getOrder(BigInteger orderId);
 
     List<Order> getOrders(Condition selector);
