+++ pet-shop-repository/pet-shop-repository-api/src/main/java/by/kachanov/shop/repository/ProductRepository.java
@@ -3,14 +3,14 @@ package by.kachanov.shop.repository;
 import by.kachanov.shop.dto.Product;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ProductRepository {
 
     void saveProduct(Product product);
 
-    Product getProduct(BigDecimal productId);
+    Product getProduct(BigInteger productId);
 
     List<Product> getProducts(Condition selector);
