+++ pet-shop-repository/pet-shop-repository-api/src/main/java/by/kachanov/shop/repository/UserRepository.java
@@ -3,12 +3,12 @@ package by.kachanov.shop.repository;
 import by.kachanov.shop.dto.User;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface UserRepository {
 
-    User getUser(BigDecimal userId);
+    User getUser(BigInteger userId);
 
     List<User> getUsers(Condition selector);
