+++ pet-shop-service/pet-shop-service-api/src/main/java/by/kachanov/shop/service/CategoryService.java
@@ -3,17 +3,17 @@ package by.kachanov.shop.service;
 import by.kachanov.shop.dto.Category;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface CategoryService {
 
     void saveCategory(Category category);
 
-    Category getCategory(BigDecimal categoryId);
+    Category getCategory(BigInteger categoryId);
 
     List<Category> getCategories(Condition selector);
 
-    void deleteCategory(BigDecimal categoryId);
+    void deleteCategory(BigInteger categoryId);
 
 }
