+++ pet-shop-service/pet-shop-service-api/src/main/java/by/kachanov/shop/service/OrderService.java
@@ -3,17 +3,17 @@ package by.kachanov.shop.service;
 import by.kachanov.shop.dto.Order;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface OrderService {
 
     void saveOrder(Order order);
 
-    Order getOrder(BigDecimal orderId);
+    Order getOrder(BigInteger orderId);
 
     List<Order> getOrders(Condition selector);
 
-    void deleteOrder(BigDecimal orderId);
+    void deleteOrder(BigInteger orderId);
 
 }
