+++ pet-shop-service/pet-shop-service-api/src/main/java/by/kachanov/shop/service/ProductService.java
@@ -3,17 +3,17 @@ package by.kachanov.shop.service;
 import by.kachanov.shop.dto.Product;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ProductService {
 
     void saveProduct(Product product);
 
-    Product getProduct(BigDecimal productId);
+    Product getProduct(BigInteger productId);
 
     List<Product> getProducts(Condition selector);
 
-    void deleteProduct(BigDecimal productId);
+    void deleteProduct(BigInteger productId);
 
 }
