+++ pet-shop-service/pet-shop-service-api/src/main/java/by/kachanov/shop/service/UserService.java
@@ -3,17 +3,17 @@ package by.kachanov.shop.service;
 import by.kachanov.shop.dto.User;
 import by.kachanov.shop.dto.condition.Condition;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 public interface UserService {
 
-    User getUser(BigDecimal userId);
+    User getUser(BigInteger userId);
 
     List<User> getUsers(Condition selector);
 
     void saveUser(User user);
 
-    void deleteUser(BigDecimal userId);
+    void deleteUser(BigInteger userId);
 
 }
